Change `count()` to type `long`

+++ src/main/java/dev/root101/spring/client/ConsumerRepoTemplate.java
@@ -80,7 +80,7 @@ public abstract class ConsumerRepoTemplate<Entity, ID> implements CRUDExternalRe
     }
 
     @Override
-    public int count() throws RuntimeException {
+    public long count() throws RuntimeException {
         return template().getForObject(urlGeneral + RESTUrlConstants.COUNT_PATH, Integer.class);
     }
